+++ src/Ci/Bamboo.php
@@ -4,6 +4,7 @@ namespace OndraM\CiDetector\Ci;
 
 use OndraM\CiDetector\CiDetector;
 use OndraM\CiDetector\Env;
+use OndraM\CiDetector\TrinaryLogic;
 
 class Bamboo extends AbstractCi
 {
@@ -15,6 +16,11 @@ class Bamboo extends AbstractCi
     public function getCiName(): string
     {
         return CiDetector::CI_BAMBOO;
+    }
+
+    public function isPullRequest(): TrinaryLogic
+    {
+        return TrinaryLogic::createFromBoolean($this->env->get('bamboo_repository_pr_key') !== false);
     }
 
     public function getBuildNumber(): string