+++ src/Ci/Circle.php
@@ -4,6 +4,7 @@ namespace OndraM\CiDetector\Ci;
 
 use OndraM\CiDetector\CiDetector;
 use OndraM\CiDetector\Env;
+use OndraM\CiDetector\TrinaryLogic;
 
 class Circle extends AbstractCi
 {
@@ -15,6 +16,11 @@ class Circle extends AbstractCi
     public function getCiName(): string
     {
         return CiDetector::CI_CIRCLE;
+    }
+
+    public function isPullRequest(): TrinaryLogic
+    {
+        return TrinaryLogic::createFromBoolean($this->env->getString('CI_PULL_REQUEST') !== '');
     }
 
     public function getBuildNumber(): string